CI: Don't run semver job for cli package
#### Problem

The semver job is failing during publish because the CLI package has no
lib target.

#### Summary of changes

I went through a few iterations on this, including a complicated cargo
metadata | jq command to see if there were any lib targets on the crate,
and then decided it would be easier to just not run the job for the cli
package.

And then while doing that, I found out that there's no way to easily
say:

* don't run a set of steps
* still mark the job as successful for later jobs that require it to be
  successful

So I added the "if" line to a bunch of steps in the semver job.