interface: Downgrade version to crates.io version
#### Problem

The version in the repo for the interface crate is at 1.0.0, but the
newest published version is 0.10. Since the publish job also bumps the
version, if we publish from here, it'll be 1.X, which is a bit
confusing.

#### Summary of changes

Downgrade the version to the one currently on crates.io which is 0.10.