deps: Upgrade to SDK v3
#### Problem

SDK v3 has been out for some time, but the sample program and CLI have
not been upgraded to it.

#### Summary of changes

Do all the upgrades! This PR also uses the split-up crates, including
SPL interface crates instead of program crates. To properly use the
interface crate, also add the token-2022 BPF program as a fixture.

Also add a `TryFrom<u32>` implementation for TransferHookError, and stop
using solana-program in the interface.